feat: move remaining simple FeedFilters to commons with ICacheProvider expansion (batch 9)
Move 4 FeedFilters that were skipped in batch 6 due to needing
collection-level cache iteration:

- RelayFeedFilter
- UserProfileMutualFeedFilter
- UserProfileAppRecommendationsFeedFilter
- WebBookmarkFeedFilter

ICacheProvider changes:
- Add notes: ICacheOperations<HexKey, Note> property
- Add addressables: ICacheOperations<Address, AddressableNote> property
- Add filterIntoSet(kind, pubKey) extension for addressable cache

IAccount changes:
- Add liveHiddenUsers(): LiveHiddenUsers for isHiddenFor() checks

All filters now use IAccount + ICacheProvider instead of Account + LocalCache.
Original files replaced with typealiases for backward compatibility.
