+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/header/apps/UserAppRecommendationsFeedViewModel.kt
@@ -23,13 +23,14 @@ package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.header.apps
 import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.vitorpamplona.amethyst.model.LocalCache
 import com.vitorpamplona.amethyst.model.User
 import com.vitorpamplona.amethyst.ui.screen.AndroidFeedViewModel
 
 @Stable
 class UserAppRecommendationsFeedViewModel(
     val user: User,
-) : AndroidFeedViewModel(UserProfileAppRecommendationsFeedFilter(user)) {
+) : AndroidFeedViewModel(UserProfileAppRecommendationsFeedFilter(user, LocalCache)) {
     class Factory(
         val user: User,
     ) : ViewModelProvider.Factory {
