+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/mutual/dal/UserProfileMutualFeedFilter.kt
@@ -20,75 +20,4 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.mutual.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.quartz.experimental.audio.header.AudioHeaderEvent
-import com.vitorpamplona.quartz.experimental.audio.track.AudioTrackEvent
-import com.vitorpamplona.quartz.experimental.interactiveStories.InteractiveStoryPrologueEvent
-import com.vitorpamplona.quartz.experimental.nipsOnNostr.NipTextEvent
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip01Core.core.AddressableEvent
-import com.vitorpamplona.quartz.nip01Core.tags.people.isTaggedUser
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip18Reposts.GenericRepostEvent
-import com.vitorpamplona.quartz.nip18Reposts.RepostEvent
-import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
-import com.vitorpamplona.quartz.nip35Torrents.TorrentEvent
-import com.vitorpamplona.quartz.nip54Wiki.WikiNoteEvent
-import com.vitorpamplona.quartz.nip84Highlights.HighlightEvent
-import com.vitorpamplona.quartz.nip99Classifieds.ClassifiedsEvent
-import com.vitorpamplona.quartz.nipA4PublicMessages.PublicMessageEvent
-
-class UserProfileMutualFeedFilter(
-    val user: User,
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + user.pubkeyHex
-
-    override fun feed(): List<Note> {
-        val notes =
-            LocalCache.notes.filterIntoSet { _, it ->
-                it !is AddressableNote && it.event !is AddressableEvent && acceptableEvent(it)
-            }
-
-        val longFormNotes =
-            LocalCache.addressables.filterIntoSet { _, it ->
-                acceptableEvent(it)
-            }
-
-        return sort(notes + longFormNotes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> = collection.filterTo(HashSet()) { acceptableEvent(it) }
-
-    fun acceptableEvent(it: Note): Boolean =
-        it.author == account.userProfile() &&
-            (
-                it.event is TextNoteEvent ||
-                    it.event is ClassifiedsEvent ||
-                    it.event is RepostEvent ||
-                    it.event is GenericRepostEvent ||
-                    it.event is LongTextNoteEvent ||
-                    it.event is WikiNoteEvent ||
-                    it.event is NipTextEvent ||
-                    it.event is ZapPollEvent ||
-                    it.event is HighlightEvent ||
-                    it.event is InteractiveStoryPrologueEvent ||
-                    it.event is AudioTrackEvent ||
-                    it.event is AudioHeaderEvent ||
-                    it.event is PublicMessageEvent ||
-                    it.event is TorrentEvent
-            ) &&
-            it.event?.isTaggedUser(user.pubkeyHex) == true
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-
-    override fun limit() = 200
-}
+typealias UserProfileMutualFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.UserProfileMutualFeedFilter
